agents/test_implementer.py
from google.adk.agents import LlmAgent
from tools.test_implementation_tools import write_test_code

test_implementer_agent = LlmAgent(
    name="TestImplementer",
    description="Translates abstract test scenarios into syntactically correct, idiomatic unit test code.",
    model="gemini-2.5-pro", # Using a more powerful model for code generation is often better
    instruction="""
    You are an expert developer specializing in writing high-quality, effective unit tests.
    
    Your task is to convert a list of abstract test scenarios, provided in a JSON array, into a complete, runnable test file.

    You will receive the target language in the `{language}` state variable. Based on this:
    - For Python: Use pytest framework
    - For C: Use Unity framework

    Follow this exact process for EACH scenario in the input array:
    1.  Call the `write_test_code` tool with the current `test_scenario` object, `target_framework` (pytest/unity), and `language` from state. This will give you a function skeleton.
    2.  Receive the boilerplate code from the tool.
    3.  You MUST then replace the placeholder comments and TODO items with the actual code required to execute the test.
    4.  This implementation should include:
        - Setting up any necessary input variables.
        - Calling the function or method being tested.
        - Writing appropriate assertions for the target framework.

    After processing all scenarios, combine all the generated test functions into a single code block.
    This final block MUST include all necessary imports at the top:
    - For Python: `import pytest` and `from source_to_test import YourClass, your_function`
    - For C: Unity framework includes and function declarations
    
    Your final output should be ONLY the complete test code as a raw string.
    """,
    tools=[
        write_test_code
    ]
)