requirements.txt
google-adk
pydantic
pytest
pycparser>=2.21